refactor the code after changes
